Optimized SAPIRObject (i.e., SAPIRFeature).

+++ src/it/cnr/isti/vir/util/L1.java
@@ -5,7 +5,7 @@ public class L1 {
 	public static int get(byte[] v1, byte[] v2) {
 		int dist = 0;		
 	    for (int i = 0; i < v1.length; i++) {
-	    	dist += Math.abs((int) v1[i] - (int) v2[i]);
+	    	dist += Math.abs(v1[i] - v2[i]);
 	    }		
 		return dist;
 	}
@@ -13,7 +13,7 @@ public class L1 {
 	public static int get(byte[] v1, byte[] v2, int maxDist) {
 		int dist = 0;
 	    for (int i = 0; i < v1.length; i++) {
-	    	dist += Math.abs((int) v1[i] - (int) v2[i]);
+	    	dist += Math.abs(v1[i] - v2[i]);
 	    	if ( dist > maxDist ) return -dist;
 	    }		
 		return dist;
@@ -22,7 +22,7 @@ public class L1 {
 	public static int get(short[] v1, short[] v2) {
 		int dist = 0;		
 	    for (int i = 0; i < v1.length; i++) {
-	    	dist += Math.abs((int) v1[i] - (int) v2[i]);
+	    	dist += Math.abs(v1[i] - v2[i]);
 	    }		
 		return dist;
 	}
@@ -30,7 +30,7 @@ public class L1 {
 	public static int get(short[] v1, short[] v2, int maxDist) {
 		int dist = 0;
 	    for (int i = 0; i < v1.length; i++) {
-	    	dist += Math.abs((int) v1[i] - (int) v2[i]);
+	    	dist += Math.abs(v1[i] - v2[i]);
 	    	if ( dist > maxDist ) return -dist;
 	    }		
 		return dist;
@@ -54,16 +54,16 @@ public class L1 {
 	}
 	
 	
-	public static int get(float[] v1, float[] v2) {
-		int dist = 0;		
+	public static float get(float[] v1, float[] v2) {
+		float dist = 0;		
 	    for (int i = 0; i < v1.length; i++) {
 	    	dist += Math.abs(v1[i] - v2[i]);
 	    }		
 		return dist;
 	}
 	
-	public static int get(float[] v1, float[] v2, int maxDist) {
-		int dist = 0;
+	public static float get(float[] v1, float[] v2, float maxDist) {
+		float dist = 0;
 	    for (int i = 0; i < v1.length; i++) {
 	    	dist += Math.abs(v1[i] - v2[i]);
 	    	if ( dist > maxDist ) return -dist;
